perform transform and coder on the correct data for padded slices
Haar only needs the values within the true slice so without modifying
the haar transform skipping the padding would be the correct thing to
do.  Testing shows there is no difference except for the picture edges
where we have yet to copy the pixel data.

The section in rate control is not done yet but const quant works fine.