add some overlap to the the transform check and loop
Transforms other than Haar need some neighbouring data to be performed
correctly.  This a part of an attempt to make it variable at runtime.
It varies with the wavelet_depth but not yet with the wavelet_type (or
wavelet_index).

Tested to have correct decoding with Haar, Legall 5_3, and Dubuc 9_7
with depths up to 4.