vc2enc: re-add 2 errors about slices sizes
Check that they are a factor of the frame size and that they are a
multiple of 2^wavelet_depth, otherwise print an error message but do not
return an error code.  This last part mnakes it easier to see the
message in vc2-send logs and allows testing without code changes.